Move referrals from browser -> components
