+++ browser/BUILD.gn
@@ -57,7 +57,6 @@ source_set("browser_process") {
   deps = [
     ":version_info",
     "//base",
-    "//brave/browser/referrals",
     "//brave/browser/tor",
     "//brave/common",
     "//brave/components/brave_rewards/browser",
@@ -109,6 +108,7 @@ source_set("browser") {
   deps = [
     "autoplay",
     "//base",
+    "//brave/components/brave_referrals/browser",
     "//brave/components/brave_shields/browser:brave_shields",
     "//brave/components/resources",
     "//brave/browser/resources:brave_extension_grit",
@@ -119,7 +119,6 @@ source_set("browser") {
     "net",
     "permissions",
     "profiles",
-    "referrals",
     "renderer_context_menu",
     "renderer_host",
     "tor",
