+++ browser/brave_browser_process_impl.cc
@@ -13,7 +13,7 @@
 #include "brave/browser/extensions/brave_tor_client_updater.h"
 #include "brave/browser/profiles/brave_profile_manager.h"
 #include "brave/browser/profile_creation_monitor.h"
-#include "brave/browser/referrals/brave_referrals_service.h"
+#include "brave/components/brave_referrals/browser/brave_referrals_service.h"
 #include "brave/components/brave_shields/browser/ad_block_service.h"
 #include "brave/components/brave_shields/browser/ad_block_regional_service.h"
 #include "brave/components/brave_shields/browser/https_everywhere_service.h"
