+++ browser/brave_local_state_prefs.cc
@@ -6,8 +6,8 @@
 
 #include "base/values.h"
 #include "brave/browser/brave_stats_updater.h"
-#include "brave/browser/referrals/brave_referrals_service.h"
 #include "brave/browser/tor/tor_profile_service.h"
+#include "brave/components/brave_referrals/browser/brave_referrals_service.h"
 #include "chrome/browser/first_run/first_run.h"
 #include "chrome/common/pref_names.h"
 #include "components/prefs/pref_registry_simple.h"
