+++ browser/net/brave_referrals_network_delegate_helper.cc
@@ -5,7 +5,7 @@
 #include "brave/browser/net/brave_referrals_network_delegate_helper.h"
 
 #include "base/values.h"
-#include "brave/browser/referrals/brave_referrals_service.h"
+#include "brave/components/brave_referrals/browser/brave_referrals_service.h"
 #include "chrome/browser/browser_process.h"
 #include "content/public/browser/browser_thread.h"
 #include "extensions/common/url_pattern.h"
