+++ components/brave_referrals/browser/brave_referrals_service.cc
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/browser/referrals/brave_referrals_service.h"
+#include "brave/components/brave_referrals/browser/brave_referrals_service.h"
 
 #include "base/environment.h"
 #include "base/files/file_path.h"
